Adding template to all 'Acteur' pages with links and redirection

+++ src/FilmoBundle/Controller/ActeurController.php
@@ -12,7 +12,7 @@ class ActeurController extends Controller
 {
     public function ajoutAction()
     {
-        $mbutton="Ajouter";
+        $mbutton = "Ajouter";
         $message = "Formulaire d'ajout d'acteur";
         $em = $this->getDoctrine()->getManager();
         $act = new Acteur();
@@ -30,7 +30,8 @@ class ActeurController extends Controller
             if ($form->isValid()) {
                 $em->persist($act);
                 $em->flush();
-                $message = "Acteur ajouté avec succés";
+                //$message = "Acteur ajouté avec succés";
+                return $this->redirectToRoute("acteur_affiche");
             }
 
         }
@@ -50,7 +51,8 @@ class ActeurController extends Controller
     }
 
     public function editAction($id)
-    {   $mbutton="Modifer";
+    {
+        $mbutton = "Modifer";
         $message = "modification d'acteur";
         $em = $this->getDoctrine()->getManager();
         $act = $em->getRepository("FilmoBundle:Acteur")->find($id);
@@ -67,7 +69,8 @@ class ActeurController extends Controller
             $form->handleRequest($request);
             if ($form->isValid()) {
                 $em->flush();
-                $message = "Acteur modifié avec succés";
+                //$message = "Acteur modifié avec succés";
+                return $this->redirectToRoute("acteur_affiche");
             }
 
         }
@@ -78,12 +81,17 @@ class ActeurController extends Controller
         ));
 
     }
+
     public function delAction($id)
     {
         $em = $this->getDoctrine()->getManager();
-        $act = $em->find("FilmoBundle:Acteur",$id);
+        $act = $em->find("FilmoBundle:Acteur", $id);
+        if (!$act) {
+            throw $this->createNotFoundException("id not found");
+        }
         $em->remove($act);
         $em->flush();
-        return new Response("Supprission effectué avec succés");
+        //return new Response("Supprission effectué avec succés");
+        return $this->redirectToRoute("acteur_affiche");
     }
 }
